backend/requirements.txt
# CounselAI Backend Dependencies

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# ── Validation & settings ─────────────────────────────────────────────────────
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic[email]==2.6.4
email-validator==2.1.1

# ── HTTP client (model service calls) ────────────────────────────────────────
httpx==0.27.0

# ── Document parsing ──────────────────────────────────────────────────────────
pypdf==4.1.0
python-docx==1.1.0

# ── AI / LLM (optional — model service) ──────────────────────────────────────
openai==1.17.0
anthropic==0.23.1

# ── Dev / testing ─────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0        # async test client
